server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
ariadne==0.26.2
arrow==1.3.0
attrs==25.3.0
bcrypt==4.3.0
beanie==1.29.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-jwt-auth==0.5.0
graphql-core==3.2.5
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
ics==0.7.2
idna==3.10
iniconfig==2.1.0
lazy-model==0.2.0
motor==3.7.0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==1.7.1
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
TatSu==5.13.1
toml==0.10.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
uvicorn==0.34.2
